Fixes #157 Create common ancestor for Gits
